feat: refactored encrypt and decrypting of data
