+++ utils/API/index.js
@@ -1,6 +1,7 @@
 export * from "./amount";
-// export * from "./decryptUserData";
-// export * from "./encryptUserData";
+export * from "./decryptData";
+export * from "./encryption_secrets";
+export * from "./encryptData";
 export * from "./response";
 export * from "./skip";
 export * from "./take";
